+++ lib/src/com/tbd/forkfront/NH_Handler.java
@@ -16,7 +16,7 @@ public interface NH_Handler
 	void clearWindow(int wid, int isRogueLevel);
 	void destroyWindow(int wid);
 	void startMenu(int wid);
-	void addMenu(int wid, int tile, int id, int acc, int groupAcc, int attr, String msg, int bSelected, int color);
+	void addMenu(int wid, int tile, long id, int acc, int groupAcc, int attr, String msg, int bSelected, int color);
 	void endMenu(int wid, String msg);
 	void selectMenu(int wid, int how);
 	void cliparound(int x, int y, int playerX, int playerY);
